Update and rename App.php to App.html
I noticed a lot of unnecessary pieces. This fork is mostly going to be about fixing that, basically making it smaller and faster for the user. There were a lot of repeated echos that didn't need to exist, as well as missing tags for the body.

There were a few tags that did not terminate.

Tip: PHP files can contain HTML. I'm also starting to wonder if we need PHP at all. The page contains nothing that needs it.

Also, we'll probably need something to replace the iframe eventually. The tag is deprecated.